# src/docs/swagger.yaml
basePath: /
definitions:
...
  main.APICarRef:
    properties:
      car_id:
        description: CarID TeslaMate cars.id
        type: integer
      car_name:
        description: CarName 车辆昵称，可为空
        type: string
    type: object
...
  main.APIDriveBatteryTrip:
    properties:
      end_battery_level:
        type: integer
      end_usable_battery_level:
        type: integer
      is_sufficiently_precise:
        type: boolean
      reduced_range:
        type: boolean
      start_battery_level:
        type: integer
      start_usable_battery_level:
        type: integer
    type: object
  main.APIDriveDetail:
    properties:
      battery_details:
        $ref: '#/definitions/main.APIDriveBatteryTrip'
      consumption_net:
        type: number
      drive_details:
        items:
          $ref: '#/definitions/main.APIDriveDetailPoint'
        type: array
      drive_id:
        type: integer
      duration_min:
        type: integer
      duration_str:
        type: string
      end_address:
        type: string
      end_date:
        type: string
      energy_consumed_net:
        type: number
      inside_temp_avg:
        type: number
      odometer_details:
        $ref: '#/definitions/main.APIOdometerTrip'
      outside_temp_avg:
        type: number
      power_max:
        type: integer
      power_min:
        type: integer
      range_ideal:
        $ref: '#/definitions/main.APIRangeTrip'
      range_rated:
        $ref: '#/definitions/main.APIRangeTrip'
      speed_avg:
        type: number
      speed_max:
        type: integer
      start_address:
        type: string
      start_date:
        type: string
    type: object
  main.APIDriveDetailBattery:
    properties:
      battery_heater:
        $ref: '#/definitions/main.NullBool'
      battery_heater_no_power:
        $ref: '#/definitions/main.NullBool'
      battery_heater_on:
        $ref: '#/definitions/main.NullBool'
      est_battery_range:
        $ref: '#/definitions/main.NullFloat64'
      ideal_battery_range:
        $ref: '#/definitions/main.NullFloat64'
      rated_battery_range:
        $ref: '#/definitions/main.NullFloat64'
    type: object
  main.APIDriveDetailClimate:
    properties:
      driver_temp_setting:
        $ref: '#/definitions/main.NullFloat64'
      fan_status:
        $ref: '#/definitions/main.NullInt64'
      inside_temp:
        $ref: '#/definitions/main.NullFloat64'
      is_climate_on:
        $ref: '#/definitions/main.NullBool'
      is_front_defroster_on:
        $ref: '#/definitions/main.NullBool'
      is_rear_defroster_on:
        $ref: '#/definitions/main.NullBool'
      outside_temp:
        $ref: '#/definitions/main.NullFloat64'
      passenger_temp_setting:
        $ref: '#/definitions/main.NullFloat64'
    type: object
  main.APIDriveDetailPoint:
    properties:
      battery_info:
        $ref: '#/definitions/main.APIDriveDetailBattery'
      battery_level:
        type: integer
      climate_info:
        $ref: '#/definitions/main.APIDriveDetailClimate'
      date:
        type: string
      detail_id:
        type: integer
      elevation:
        $ref: '#/definitions/main.NullInt64'
      latitude:
        type: number
      longitude:
        type: number
      odometer:
        type: number
      power:
        type: integer
      speed:
        type: integer
      usable_battery_level:
        $ref: '#/definitions/main.NullInt64'
    type: object
...
  main.APIOdometerTrip:
    properties:
      odometer_distance:
        type: number
      odometer_end:
        type: number
      odometer_start:
        type: number
    type: object
...
  main.APIRangeTrip:
    properties:
      end_range:
        type: number
      range_diff:
        type: number
      start_range:
        type: number
    type: object
...
  main.APIUnitsLengthTemp:
    properties:
      unit_of_length:
        description: UnitsLength km 或 mi
        type: string
      unit_of_temperature:
        description: UnitsTemperature C 或 F
        type: string
    type: object
...
  main.NullBool:
    properties:
      bool:
        type: boolean
      valid:
        description: Valid is true if Bool is not NULL
        type: boolean
    type: object
  main.NullFloat64:
    properties:
      float64:
        type: number
      valid:
        description: Valid is true if Float64 is not NULL
        type: boolean
    type: object
  main.NullInt64:
    properties:
      int64:
        type: integer
      valid:
        description: Valid is true if Int64 is not NULL
        type: boolean
    type: object
  main.RespAPIError:
    properties:
      error:
        description: Error 错误说明
        type: string
    type: object
...
  main.RespDriveDetail:
    properties:
      data:
        $ref: '#/definitions/main.RespDriveDetailData'
    type: object
  main.RespDriveDetailData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      drive:
        $ref: '#/definitions/main.APIDriveDetail'
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 与 TeslaMate 采集数据对齐的 JSON API；看板对照见仓库 `执行计划.md`。
  license:
    name: MIT
  title: TeslaMate API
  version: "1.0"
paths:
...
  /api/v1/cars/{CarID}/command:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
  /api/v1/cars/{CarID}/command/{Command}:
    post:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 指令名（POST /command/:Command）
        in: path
        name: Command
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
...
  /api/v1/cars/{CarID}/drives/{DriveID}:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 行程 ID
        in: path
        name: DriveID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespDriveDetail'
      summary: 行程详情
      tags:
      - drives
...
  /api/v1/cars/{CarID}/wake_up:
    post:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: POST /command/*、/wake_up：Tesla API 原始 JSON
          schema:
            items:
              type: integer
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.RespAPIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.RespAPIError'
      summary: 车辆指令 / 唤醒
      tags:
      - commands
...
schemes:
- http
swagger: "2.0"